This PR makes PostgreSQL non-blocking and play nice with Libvenice:

- Uses async connection
- Uses async querying

@Danappelxx please review 👍