@Danappelxx 2016-12-17
Two things lead me to think that we're not polling _quite_ right:

1. In my testing, the blocking version was actually faster than this nonblocking version. In theory, this is not supposed to happen. I tested using `wrk -d 10 -t 10 -c 100 http://localhost:8080/api/v1/posts` on an endpoint which is 100% database io.

without this pr:
```
> wrk -d 10 -t 10 -c 100 http://localhost:8080/api/v1/posts
Running 10s test @ http://localhost:8080/api/v1/posts
  10 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    44.86ms    3.71ms 101.55ms   87.31%
    Req/Sec   223.54     22.24   300.00     75.20%
  22326 requests in 10.05s, 8.84MB read
Requests/sec:   2221.08
Transfer/sec:      0.88MB
```

with this pr:
```
> wrk -d 10 -t 10 -c 100 http://localhost:8080/api/v1/posts
Running 10s test @ http://localhost:8080/api/v1/posts
  10 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     3.06ms    5.66ms  82.92ms   87.77%
    Req/Sec   670.91    280.70     1.16k    66.05%
  10877 requests in 10.08s, 3.56MB read
  Socket errors: connect 0, read 2360, write 0, timeout 0
  Non-2xx or 3xx responses: 2459
Requests/sec:   1079.14
Transfer/sec:    361.59KB
```

2. In the logs for this test, I found 2425 occurrences of `Error: another command is already in progress`. This is probably a symptom of the same issue :)