@robertjpayne 2016-12-17
@Danappelxx as per the performance your tests are re-using a connection I assume? This is why it appears to perform slower.

Before a single request will execute the full SQL command before any other request in the entire process could continue thus you sort of have an implicit lock on any SQL execution.

Now when libpq is waiting for IO another coroutine will pick up a new request and start to try and use the same Postgres connection, Postgres is half way through executing another SQL statement and hence you get the errors.

For this async stuff to work we really need a connection pool that a request checks a connection out and when finished it puts it back.