@Danappelxx 2016-12-23
Did some more testing and I'm satisfied! Works great with a connection pool as you suggested. Just a little bit of cleanup and its good to go.